+++ changelog.md
@@ -1,3 +1,4 @@
-v0.0.5:
-- extracted some logic from `reg_cuts` to `parse_keycomb`
-- shortcuts syntax updated
+v0.1.0:
+- new shortcuts config syntax (see wiki)
+- shortcuts now may switch workspaces
+- workspaces! see wiki